+++ CHANGELOG.md
@@ -12,8 +12,7 @@ Each version section may have have subsections for: _Added_, _Changed_, _Removed
 
 ## [1.18.68]
 ### Added
-- Source factors can now be added to the embeddings instead of concatenated (`--source-factors-combine sum`)
-- Options are 'sum' (the default) or 'concat"
+- Source factors can now be added to the embeddings instead of concatenated with `--source-factors-combine sum` (default: concat)
 
 ## [1.18.67]
 ### Added